RW-41: add uniq id to offline plant, implement total count function on db, rename files
