+++ upload-app/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@mui/base": "^5.0.0-beta.33",
     "@tanstack/react-query": "^5.17.15",
+    "@types/uuid": "^9.0.8",
     "axios": "^1.6.7",
     "clsx": "^2.1.0",
     "date-fns": "^3.3.1",
@@ -23,6 +24,7 @@
     "react-intl": "^6.6.1",
     "react-router-dom": "^6.21.3",
     "react-spinners": "^0.13.8",
+    "uuid": "^9.0.1",
     "zustand": "^4.5.0"
   },
   "devDependencies": {
