+++ upload-app/src/modules/plants/components/WaitingSubmissions.tsx
@@ -1,14 +1,13 @@
 import { LoaderBox } from 'common/components';
 import { useFmtMsg } from 'modules/intl';
-import { loadNewPlants } from 'modules/offline-db';
+import { OfflinePlant, loadNewPlants } from 'modules/offline';
 import { useEffect, useState } from 'react';
-import { NewPlant } from '..';
 import { WaitingPlantTile } from './WaitingPlantTile';
 
 export const WaitingSubmissions: React.FC = () => {
   const fmtMsg = useFmtMsg();
   const [isLoading, setIsLoading] = useState(true);
-  const [plants, setPlants] = useState<NewPlant[]>();
+  const [plants, setPlants] = useState<OfflinePlant[]>();
 
   useEffect(() => {
     loadNewPlants().then((plants) => {
@@ -21,7 +20,7 @@ export const WaitingSubmissions: React.FC = () => {
     <div className='space-y-2.5'>
       <LoaderBox visible={isLoading} />
       {plants?.map((plant) => (
-        <WaitingPlantTile plant={plant} />
+        <WaitingPlantTile key={plant.id} plant={plant.plant} />
       ))}
       {plants?.length === 0 && (
         <div className='text-center'>
